a contained item _might_ be not an int

+++ schema.js
@@ -141,7 +141,7 @@ module.exports = `
   type ContainedItem {
     item: Item!
     count: Float!
-    quantity: Int!
+    quantity: Float!
   }
 
   type Barter {
